Simplificación parte back con inclusión de ORM

+++ couples-back/src/main/java/couples/repositories/ScoreRepository.java
@@ -1,5 +1,4 @@
 package couples.repositories;
-
 import java.util.List;
 
 import org.springframework.data.jpa.repository.Query;
@@ -7,7 +6,7 @@ import org.springframework.data.repository.*;
 
 import couples.model.Score;
 
-public interface ScoreRepository extends CrudRepository<Score, Integer>  {
+public interface ScoreRepository extends CrudRepository<Score, Long>  {
 	
 	
 	//@Query("SELECT s , u FROM Score s, User u WHERE s.email=u.email")
